backend/requirements.txt
absl-py==0.9.0
aioredis==1.3.1
appdirs==1.4.3
as==0.1
asgiref==3.2.5
astroid==2.3.3
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
beautifulsoup4==4.6.0
blis==0.7.3
blist==1.3.6
boto3==1.16.30
botocore==1.19.30
bs4==0.0.1
cached-property==1.5.1
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.1
colorama==0.4.3
colored==1.4.2
conda==4.3.16
configparser==5.0.0
constantly==15.1.0
crayons==0.3.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.4
daphne==2.4.1
dataclasses==0.6
decorator==4.4.2
distlib==0.3.0
Django==3.0.4
django-channels==0.7.0
docutils==0.16
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.3.1/en_core_web_md-2.3.1.tar.gz
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
future==0.18.2
gast==0.3.3
gdown==3.10.3
google-api-core==1.20.0
google-api-python-client==1.10.0
google-auth==1.17.2
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
googledrivedownloader==0.4
grpcio==1.30.0
gunicorn==20.0.4
h5py==2.10.0
hiredis==1.0.1
httplib2==0.18.1
hyperlink==19.0.0
idna==2.9
imageio==2.9.0
incremental==17.5.0
iniconfig==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
langdetect==1.0.8
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.5.2
lyricsgenius==1.8.6
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
msgpack==0.6.2
murmurhash==1.0.4
networkx==2.5
nltk==3.5
numpy==1.18.2
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.7
pandas==1.0.3
pdf2image==1.13.1
pdfminer.six==20200104
pdfplumber==0.5.21
pid==3.0.0
Pillow==7.2.0
plac==1.1.3
pluggy==0.13.1
preshed==3.0.4
protobuf==3.12.2
psycopg2-binary==2.8.6
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
pycryptodome==3.9.8
PyDrive==1.3.1
PyHamcrest==2.0.2
pylint==2.4.4
pymongo==3.10.1
pynder==0.0.13
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
PySocks==1.7.1
pytesseract==0.3.4
pytest==6.1.2
python-daemon==2.2.4
python-dateutil==2.8.1
python-documentcloud==1.1.1
python-dotenv==0.15.0
pytorch-pretrained-bert==0.6.2
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
rake-nltk==1.0.4
redis==3.4.1
regex==2020.7.14
requests==2.23.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
robobrowser==0.5.3
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
sacremoses==0.0.43
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
sentencepiece==0.1.94
service==0.6.0
service-identity==18.1.0
setproctitle==1.1.10
six==1.14.0
sklearn==0.0
sortedcontainers==2.2.2
soupsieve==2.0
spacy==2.3.4
spotipy==2.13.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
srsly==1.0.4
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
textblob==0.15.3
thinc==7.4.3
threadpoolctl==2.1.0
tifffile==2020.11.26
tokenizers==0.8.0rc4
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
tqdm==4.45.0
transformers==3.0.0
Twisted==19.10.0
txaio==20.1.1
typing-extensions==3.7.4.3
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.8
validators==0.18.1
virtualenv==20.0.10
Wand==0.6.2
wasabi==0.8.0
webdriver-manager==2.5.2
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.3.3
xgboost==1.1.1
zmq==0.0.0
zope.interface==4.7.2
